requirements.txt
telethon
python-dotenv
litellm
prompt_toolkit